test: move test file to more appropriate location
